requirements.txt
# Clinical Intent Extraction & 3D Pipeline
openai>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-openrouter
langchain-groq>=0.2.0
python-dotenv>=1.0.0

# FastAPI backend and imaging stack
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
nibabel>=5.2.0
SimpleITK>=2.3.0
pillow>=10.0.0
scikit-image>=0.24.0
trimesh>=4.4.0
gradio>=4.0.0
httpx>=0.27.0
